+++ app/src/main/java/dika/wardani/domain/Collection.kt
@@ -1,3 +1,3 @@
 package dika.wardani.domain
 
-data class Collection(var id: Long, var name: String)
+data class Collection(var id: Int, var name: String)
